Created Water Gauge
visual representation of player's water level is now complete and added
to the 2d side scroller prefab. it's a child of the player game object,
and resizes dynamically during gameplay. new script was made to handle
resizing water gauge and additions to the attributescript were made to
utilize the new watergauge script appropriately